Add VolumeCreate call
It appears that golang JSON handling requires the struct members to
be public to marshal/unmarshal.  We need the class and the plugin_data
for the IPC to work, thus I've also made them public, but ideally they
would be private but still be marshalled/unmarshalled.  I tried using
a custom class type with it's own custome marshal/unmarshal, but ran into
the issue of the marshal not getting called.  Will need to revisit if
there is a better way to handle this without making these two struct
members public.

Signed-off-by: Tony Asleson <tasleson@redhat.com>