Add custom MarshalJSON for lsm object types
LibStorageMgmt tags each 'class' with a member key 'class' with a value
that matches the expected class type.  This was done so that the python
code could very easily de-serialize the object into the correct type.  Thus
for all supported languages we need to provide this meta data.  The default
json handling in golang only serializes/de-serialize public fields, thus
<object>.Class is public and thus changeable by user.  To avoid issues, when
we marshal one of these classes we will simply write the object type into
<object>.Class to ensure correct behavior.  Perhaps there is a simpler way
to achieve this, but attempts to make a custom type like:

type LsmClassVolume string

With a custom Marshal/UnMarshal resulted in a json.Marshal error:
&json.MarshalerError{Type:(*reflect.rtype)(0x72a680), Err:(*json.SyntaxError)(0xc00000ebe0)

which I was unable to correct, but I might simply be doing something wrong.

Signed-off-by: Tony Asleson <tasleson@redhat.com>